fix: make address navigation switch to debugger tab automatically
- Refactor App component to separate AppContent that can use navigation context
- Subscribe to navigation events in AppContent to switch tabs when navigating
- Navigation from Inspector view now properly switches to Debugger tab
- Fixes issue where clicking 'View in Memory Editor' or 'View in Disassembly' didn't show the target view
